Keep CLI parser under mutation site threshold
