+++ packages/core/src/cli/parse.ts
@@ -1,37 +1,13 @@
-export type CliCommandName = 'help' | 'index' | 'plugins' | 'setup' | 'status';
-export type PluginsCommandAction = 'disable' | 'enable' | 'help' | 'list' | 'register';
+import { isHelpCommandName } from './parseHelp';
+import { parsePluginsCommand } from './parsePlugins';
+import { parseWorkspaceCommand } from './parseWorkspace';
+import type { CliCommand } from './parseTypes';
 
-export interface CliCommand {
-  name: CliCommandName;
-  action?: PluginsCommandAction;
-  packageName?: string;
-  workspacePath?: string;
-}
-
-function parsePluginsCommand(argv: string[]): CliCommand {
-  const [action, packageName, workspacePath] = argv;
-
-  switch (action) {
-    case 'register':
-      return { name: 'plugins', action, packageName };
-    case 'disable':
-    case 'enable':
-      return { name: 'plugins', action, packageName, workspacePath };
-    case 'list':
-      return { name: 'plugins', action, workspacePath: packageName };
-    default:
-      return { name: 'plugins', action: 'help' };
-  }
-}
-
-function isHelpCommandName(name: string | undefined): boolean {
-  return name === undefined || name === 'help' || name === '--help' || name === '-h';
-}
-
-function parseWorkspaceCommand(name: 'index' | 'status', argv: string[]): CliCommand {
-  const [workspacePath] = argv;
-  return workspacePath ? { name, workspacePath } : { name };
-}
+export type {
+  CliCommand,
+  CliCommandName,
+  PluginsCommandAction,
+} from './parseTypes';
 
 export function parseCliCommand(argv: string[]): CliCommand {
   const [name, ...rest] = argv;
