+++ packages/core/src/cli/parseHelp.ts
@@ -1,0 +1,3 @@
+export function isHelpCommandName(name: string | undefined): boolean {
+  return name === undefined || name === 'help' || name === '--help' || name === '-h';
+}
