+++ packages/core/src/cli/parseWorkspace.ts
@@ -1,0 +1,6 @@
+import type { CliCommand } from './parseTypes';
+
+export function parseWorkspaceCommand(name: 'index' | 'status', argv: string[]): CliCommand {
+  const [workspacePath] = argv;
+  return workspacePath ? { name, workspacePath } : { name };
+}
